feat(attendance): add GET module attendance endpoint
