+++ packages/backend/src/routes/modules.routes.ts
@@ -1,6 +1,7 @@
 import { Router } from "express";
 import { requireAuth } from "../middleware/auth.middleware";
 import { requireRole } from "../middleware/role.middleware";
+import { getModuleAttendance } from "../controllers/attendance.controller";
 import {
   createModuleController,
   listMyModulesController,
@@ -14,6 +15,7 @@ const router = Router();
 
 router.get("/", requireAuth, listMyModulesController);
 router.get("/:id", requireAuth, getModuleInfoController);
+router.get("/:modules/attendance", requireAuth, getModuleAttendance);
 router.post("/", requireAuth, requireRole("COACH"), createModuleController);
 router.post("/join", requireAuth, requireRole("COACH", "ATHLETE", "PARENT"), joinModuleController);
 router.patch("/:id", requireAuth, updateModuleController);
